+++ src/lib/getContainer.ts
@@ -4,7 +4,6 @@ import settings from '../config/settings';
 import { Redis } from 'ioredis';
 import buildRedisConnection from '../utils/buildRedisConnection';
 import { Logger } from 'winston';
-import StatsdClient from 'statsd-client';
 import { Blockchain } from './Blockchain';
 import { ProjectAuthUtils } from '../services/ProjectAuthUtils';
 import LockRepository from '../repositories/LockRepository';
@@ -12,7 +11,6 @@ import { BlockchainRepository } from '../repositories/BlockchainRepository';
 import { ChainContractRepository } from '../repositories/ChainContractRepository';
 import { ManagementClient } from 'auth0';
 import { initAuth0ManagementClient } from '../config/initAuth0ManagementClient';
-import { getStatsdClient } from './getStatsdClient';
 import { getLogger } from './getLogger';
 
 export function getContainer(): Container {
@@ -31,11 +29,6 @@ export function getContainer(): Container {
   container
     .bind<Redis>('Redis')
     .toDynamicValue((ctx) => buildRedisConnection(ctx.container.get<Settings>('Settings').redis))
-    .inSingletonScope();
-
-  container
-    .bind<StatsdClient>('StatsdClient')
-    .toDynamicValue(() => getStatsdClient())
     .inSingletonScope();
 
   container