+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- in `testBalanceThreshold` support case where `balance` is null/undefined (solana case) 
+
+### Updated
+- docker node to v18
+
+### Removed
+- remove `newrelic` and `statsd`
 
 ## [5.8.0] - 2023-05-29
 ### Removed